Tutorial_1_Docker/Ejercicio_1/NestorApp/requirements.txt
flask==1.1.2
slackclient==2.9.1
pymongo==3.9.0
